refactor: move error_expression-about to expr
